Create After-pseudo-class.html
It works similar to ::before pseudo-element, but it inserts the content after the content of the element. It is used to add something after the specific part of an element. Generally, it is used with the content property.

It also allows us to add regular strings or images after the content.