Create First-line-pseudo-class.html
It is similar to the ::first-letter pseudo-element, but it affects the entire line. It adds the special effects to the first line of the text. It supports the following CSS properties:

Color properties (such as color)
Font properties (such as font-style, font-family, font-size, font-color, and many more).
Background properties (such as background-color, background-repeat, background-image, and background-position).
Other properties are word-spacing, letter-spacing, line-height, vertical-align, text-transform, text-decoration.